+++ src/components/Checkbox/Checkbox.visual.tsx
@@ -1,13 +1,17 @@
 import * as React from 'react';
 import { snap, story, visualize } from 'storybook-snapper';
 import { VisualTestContainer } from '../../../test/visual/VisualTestContainer';
-import { Checkbox, CheckboxTheme } from './';
+import { Checkbox, CheckboxTheme, CheckboxProps } from './';
 
-class CheckboxVisual extends React.Component<any> {
+class CheckboxVisual extends React.Component<Partial<CheckboxProps>> {
   render() {
     return (
       <VisualTestContainer>
-        <Checkbox {...this.props} label="Amazing" onChange={() => {}} />
+        <Checkbox
+          {...this.props}
+          label={this.props.label || 'Amazing'}
+          onChange={() => {}}
+        />
       </VisualTestContainer>
     );
   }
@@ -56,6 +60,24 @@ visualize('Checkbox', () => {
     snap(
       `${CheckboxTheme.Box} / suffix`,
       <CheckboxVisual theme={CheckboxTheme.Box} suffix="$50,000" />,
+    );
+  });
+
+  story('long', () => {
+    snap(
+      `${CheckboxTheme.Box} / long`,
+      <CheckboxVisual
+        theme={CheckboxTheme.Box}
+        label="It is a period of civil war. Rebel spaceships, striking from a hidden base, have won their first victory against the evil Galactic Empire. During the battle, Rebel spies managed to steal secret plans to the Empire's ultimate weapon, the DEATH STAR, an armored space station with enough power to destroy an entire planet."
+      />,
+    );
+    snap(
+      `${CheckboxTheme.Box} / long / suffix`,
+      <CheckboxVisual
+        theme={CheckboxTheme.Box}
+        suffix="$50,000"
+        label="It is a period of civil war. Rebel spaceships, striking from a hidden base, have won their first victory against the evil Galactic Empire. During the battle, Rebel spies managed to steal secret plans to the Empire's ultimate weapon, the DEATH STAR, an armored space station with enough power to destroy an entire planet."
+      />,
     );
   });
 });