+++ src/components/FloatingDropdown/FloatingDropdown.tsx
@@ -15,6 +15,7 @@ import { Option, OptionFactory } from 'wix-ui-core/dropdown-option';
  * https://github.com/wix/wix-ui-tpa/pull/387 (or merged into a single component)
  */
 import '../Dropdown/Dropdown.st.css';
+import classnames from 'classnames';
 
 export interface FloatingDropdownProps extends TPAComponentProps {
   /** Defines a string value that labels the current element. Optional. */
@@ -43,10 +44,15 @@ export interface FloatingDropdownProps extends TPAComponentProps {
   onExpandedChange?(isExpanded: boolean): void;
   name?: string;
   id?: string;
+  /** Sets the display type of the component to be block if true */
+  displayBlock?: boolean;
 }
 
 type DefaultProps = Required<
-  Pick<FloatingDropdownProps, 'disabled' | 'options' | 'placement'>
+  Pick<
+    FloatingDropdownProps,
+    'disabled' | 'options' | 'placement' | 'displayBlock'
+  >
 >;
 
 /** Dropdown component for sort. */
@@ -56,6 +62,7 @@ export class FloatingDropdown extends React.Component<FloatingDropdownProps> {
     disabled: false,
     options: [],
     placement: 'bottom',
+    displayBlock: false,
   };
 
   _onSelect = (selectedOption: FloatingDropdownOptionProps) => {
@@ -72,7 +79,7 @@ export class FloatingDropdown extends React.Component<FloatingDropdownProps> {
       return;
     }
     const selectedOption = this.props.options.find(
-      ({ id }) => selectedCoreOption.id === id,
+      ({ id }) => selectedCoreOption.id === id
     );
     this._onSelect(selectedOption);
   };
@@ -80,7 +87,7 @@ export class FloatingDropdown extends React.Component<FloatingDropdownProps> {
   _generateCoreOptions() {
     const { options } = this.props;
 
-    const coreOptions = options.map(option =>
+    const coreOptions = options.map((option) =>
       option.divider
         ? OptionFactory.createDivider({ className: classes.optionDivider })
         : OptionFactory.create({
@@ -88,7 +95,7 @@ export class FloatingDropdown extends React.Component<FloatingDropdownProps> {
             render: () => (
               <DropdownOption className={classes.option} {...option} />
             ),
-          }),
+          })
     );
 
     return {
@@ -111,6 +118,7 @@ export class FloatingDropdown extends React.Component<FloatingDropdownProps> {
       placement,
       appendTo,
       onExpandedChange,
+      displayBlock,
     } = this.props;
     const { coreOptions } = this._generateCoreOptions();
 
@@ -128,6 +136,7 @@ export class FloatingDropdown extends React.Component<FloatingDropdownProps> {
         options={options}
         id={id}
         onChange={this._onSelect}
+        displayBlock={displayBlock}
       />
     );
 
@@ -166,8 +175,9 @@ export class FloatingDropdown extends React.Component<FloatingDropdownProps> {
             className={st(
               classes.root,
               { mobile },
-              classes.overrideStyleParams,
-              this.props.className,
+              classnames(classes.overrideStyleParams, this.props.className, {
+                [classes.displayInline]: !this.props.displayBlock,
+              })
             )}
           >
             {this._getContent(mobile)}