+++ src/components/Image/Image.tsx
@@ -1,14 +1,11 @@
 import * as React from 'react';
 import { Image as CoreImage } from 'wix-ui-core/image';
-import {
-  MediaImage,
-  MediaPlatformItem as MediaItemProps,
-} from 'wix-ui-core/media-image';
+import { MediaImage, MediaImageScaling } from 'wix-ui-core/media-image';
 import { TPAComponentProps } from '../../types';
 import { classes, st } from './Image.st.css';
 
 export interface ImageProps extends TPAComponentProps {
-  /** The full URL of the source */
+  /** The source could be any absolute full URL or a relative URI of a media platform item */
   src?: string;
   /** The intrinsic width of the image in pixels */
   width?: number;
@@ -20,41 +17,36 @@ export interface ImageProps extends TPAComponentProps {
   onLoad?: React.EventHandler<React.SyntheticEvent>;
   /** A callback to be called if error occurs while loading */
   onError?: React.EventHandler<React.SyntheticEvent>;
-  /** Exposing the MediaPlatformItem API as specified in https://wix-wix-ui-core.surge.sh/?activeTab=Usage&path=%2Fstory%2Fcomponents--media-image */
-  mediaItem?: MediaItemProps;
 }
 
-/** Image is a component to literally display an image - whether a regular with full path or an item from the media manager */
+/** Image is a component to literally display an image - whether an absolute with full URL or a media platform item with relative URI */
 export class Image extends React.Component<ImageProps> {
   static displayName = 'Image';
 
   render() {
-    const {
-      className,
-      src,
-      width,
-      height,
-      mediaItem,
-      ...imageProps
-    } = this.props;
+    const { className, src, width, height, ...imageProps } = this.props;
     const dimensions = { width, height };
+
+    const isAbsoluteUrl = src.match('^https?://');
 
     return (
       <div
         className={st(classes.root, className)}
         data-hook={this.props['data-hook']}
       >
-        {mediaItem ? (
-          <MediaImage
-            {...imageProps}
-            {...dimensions}
-            mediaPlatformItem={mediaItem}
-          />
-        ) : (
+        {isAbsoluteUrl ? (
           <CoreImage
             {...imageProps}
             nativeProps={{ ...dimensions }}
             src={src}
+          />
+        ) : (
+          <MediaImage
+            {...imageProps}
+            mediaPlatformItem={{
+              uri: src,
+              ...dimensions,
+            }}
           />
         )}
       </div>