+++ src/components/Image/docs/examples.ts
@@ -1,6 +1,6 @@
 export const importExample = `import { Image } from 'wix-ui-tpa/Image';`;
 
-export const regularImageExample = `
+export const absoluteUrlExample = `
 <Image
   src="https://m.media-amazon.com/images/M/MV5BZGMwOGIwZjUtOWQ1OS00YWRjLWJmZGMtN2Y1OWQ3ZDYwYTM3XkEyXkFqcGdeQXVyNzU1NzE3NTg@._V1_.jpg"
   width="300"
@@ -9,11 +9,9 @@ export const regularImageExample = `
 />
 `;
 
-export const mediaImageExample = `
+export const relativeUriExample = `
 <Image
-  mediaItem={{
-    uri: 'c5f754_dd75514d14fa4057b4f4a6cc8ce7add3~mv2.jpg',
-  }}
+  src="c5f754_dd75514d14fa4057b4f4a6cc8ce7add3~mv2.jpg"
   width="300"
   height="250"
   alt="Garfield smiles and puts his hand over chest"