+++ src/components/RadioButton/docs/examples.ts
@@ -1,11 +1,11 @@
 export const importExample = `import { RadioButton } from 'wix-ui-tpa/RadioButton';`;
 
 export const uncheckedExample = `
-<RadioButton value={'Unchecked'} onChange={val => console.log(val)} label="Unchecked"/>
+<RadioButton value={'Unchecked'} withFocusRing onChange={val => console.log(val)} label="Unchecked"/>
 `;
 
 export const checkedExample = `
-<RadioButton value={'Checked'} onChange={val => console.log(val)} checked label="Checked" />
+<RadioButton value={'Checked'} withFocusRing onChange={val => console.log(val)} checked label="Checked" />
 `;
 
 export const disabledExample = `
@@ -13,7 +13,7 @@ export const disabledExample = `
 `;
 
 export const boxExample = `
-<RadioButton value={'Checked'}  checked theme='box' onChange={val => console.log(val)}  label="Checked" />
+<RadioButton value={'Checked'} withFocusRing checked theme='box' onChange={val => console.log(val)}  label="Checked" />
 `;
 
 export const boxExampleDisabled = `
@@ -25,9 +25,13 @@ export const boxExampleDisabledChecked = `
 `;
 
 export const suffixExample = `
-<RadioButton value={'Checked'} suffix="$" checked  onChange={val => console.log(val)}  label="Checked" />
+<RadioButton value={'Checked'} withFocusRing suffix="$" checked  onChange={val => console.log(val)}  label="Checked" />
 `;
 
 export const errorExample = `
-<RadioButton value={'Checked'} error checked theme='box' onChange={val => console.log(val)}  label="Checked" />
+<RadioButton value={'Checked'} withFocusRing error checked theme='box' onChange={val => console.log(val)}  label="Checked" />
+`;
+
+export const withChildrenExample = `
+<RadioButton value={'Checked'} withFocusRing checked onChange={val => console.log(val)} ><span>With Children</span></RadioButton>
 `;