+++ src/components/index.ts
@@ -47,6 +47,7 @@ export { Dialog } from './Dialog';
 export { AddItem } from './AddItem';
 export { ProgressBar } from './ProgressBar';
 export { Image } from './Image';
+export { DatePicker } from './DatePicker';
 export { DatePickerInput } from './DatePickerInput';
 export { SimpleDatePicker } from './SimpleDatePicker';
 export { ThemedDatePicker } from './ThemedDatePicker';